Updated scripts that obtain data from other sources into a single script
